+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Rework copyright headers
 
+### Fixed
+
+- Cleanup copy right headers. Update to dart 3.13. Auto fixes.
+
 ## 1.1.7 - 2026-08-11
 
 ### Added