+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixed
 
 - Cleanup copy right headers. Update to dart 3.13. Auto fixes.
+- Cleanup copy right headers. Update to dart 3.13. Auto fixes. Setup quick-check pipeline.
 
 ## 1.1.7 - 2026-08-11