+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.2.0 - 2026-08-13
+
+### Changed
+
+- Rework copyright headers
+
+### Fixed
+
+- Cleanup copy right headers. Update to dart 3.13. Auto fixes.
+- Cleanup copy right headers. Update to dart 3.13. Auto fixes. Setup quick-check pipeline.
+
 ## 1.1.7 - 2026-08-11
 
 ### Added